Lizenzverwaltung: Lizenz-Editor ins Modul verschieben
