+++ src/renderer/modules/lizenzverwaltung/index.js
@@ -1,4 +1,4 @@
-import { LicenseAdminScreen, LIZENZVERWALTUNG_WORK_SCREEN_ID } from "./screens/index.js";
+import { LicenseAdminScreen, createLicenseEditorSection, LIZENZVERWALTUNG_WORK_SCREEN_ID } from "./screens/index.js";
 
 export const LIZENZVERWALTUNG_MODULE_ID = "lizenzverwaltung";
 export const LIZENZVERWALTUNG_MODULE_LABEL = "Lizenzverwaltung";
@@ -18,5 +18,5 @@ export function getLizenzverwaltungModuleEntry() {
   });
 }
 
-export { LicenseAdminScreen, LIZENZVERWALTUNG_WORK_SCREEN_ID };
+export { LicenseAdminScreen, createLicenseEditorSection, LIZENZVERWALTUNG_WORK_SCREEN_ID };
 export * from "./screens/index.js";
