+++ src/renderer/modules/lizenzverwaltung/screens/index.js
@@ -1,3 +1,4 @@
 export { default as LicenseAdminScreen } from "./LicenseAdminScreen.js";
+export { createLicenseEditorSection } from "./createLicenseEditorSection.js";
 
 export const LIZENZVERWALTUNG_WORK_SCREEN_ID = "licenseAdmin";
